Move the `/pkg/assets/dynamic.aferoFS` implem into its own file
At the moment the `fs.go` file contains 3 differents implementation.
Separate them into differents files would improve the readability.
